Backend/requirements.txt
fastapi
uvicorn
python-dotenv
databases
sqlalchemy
asyncpg
pydantic
supabase
httpx
